@@ -4,12 +4,15 @@ import (
 	"fmt"
 	"net/http"
 	"sync"
+	"time"
 )
 
 // ProviderSetup manages multiple mock RPC servers
 type ProviderSetup struct {
 	servers []httpServer
 	wg      sync.WaitGroup
+	errChan chan error
+	done    chan struct{}
 }
 
 type httpServer interface {
@@ -21,6 +24,8 @@ type httpServer interface {
 func NewProviderSetup() *ProviderSetup {
 	return &ProviderSetup{
 		servers: make([]httpServer, 0),
+		errChan: make(chan error, 10), // Buffered channel to prevent blocking
+		done:    make(chan struct{}),
 	}
 }
 
@@ -51,7 +56,13 @@ func (p *ProviderSetup) StartAll() error {
 		p.wg.Add(1)
 		go func(s httpServer) {
 			defer p.wg.Done()
-			s.Start()
+			if err := s.Start(); err != nil {
+				select {
+				case p.errChan <- err:
+				case <-p.done:
+					// Channel is closed, ignore error
+				}
+			}
 		}(server)
 	}
 	return nil
@@ -59,13 +70,36 @@ func (p *ProviderSetup) StartAll() error {
 
 // StopAll stops all mock providers
 func (p *ProviderSetup) StopAll() error {
+	// Signal that we're shutting down
+	close(p.done)
+
+	// Stop all servers
 	for _, server := range p.servers {
 		if err := server.Stop(); err != nil {
 			return err
 		}
 	}
-	p.wg.Wait()
-	return nil
+
+	// Wait for all goroutines to complete with timeout
+	done := make(chan struct{})
+	go func() {
+		p.wg.Wait()
+		close(done)
+	}()
+
+	select {
+	case <-done:
+		// Close error channel after all goroutines are done
+		close(p.errChan)
+		return nil
+	case <-time.After(5 * time.Second):
+		return fmt.Errorf("timeout waiting for servers to stop")
+	}
+}
+
+// Close ensures all resources are properly cleaned up
+func (p *ProviderSetup) Close() error {
+	return p.StopAll()
 }
 
 // http404Server wraps http.Server to implement httpServer interface
